+++ src/Repository/UserPlaylistRepository.php
@@ -20,29 +20,4 @@ class UserPlaylistRepository extends ServiceEntityRepository
     {
         parent::__construct($registry, UserPlaylist::class);
     }
-
-//    /**
-//     * @return UserPlaylist[] Returns an array of UserPlaylist objects
-//     */
-//    public function findByExampleField($value): array
-//    {
-//        return $this->createQueryBuilder('u')
-//            ->andWhere('u.exampleField = :val')
-//            ->setParameter('val', $value)
-//            ->orderBy('u.id', 'ASC')
-//            ->setMaxResults(10)
-//            ->getQuery()
-//            ->getResult()
-//        ;
-//    }
-
-//    public function findOneBySomeField($value): ?UserPlaylist
-//    {
-//        return $this->createQueryBuilder('u')
-//            ->andWhere('u.exampleField = :val')
-//            ->setParameter('val', $value)
-//            ->getQuery()
-//            ->getOneOrNullResult()
-//        ;
-//    }
 }
