+++ templates/video/player.html.twig
@@ -24,7 +24,7 @@
 								{% for playlist in playlistsIncVideo %}
 									<li class="playlistsIncVideo">
 										<a href="{{ path('playlist_show', {'id': playlist.id}) }}">{{ playlist.label }}</a>
-										<a href="{{ path('video_remove_playlist', {'playlistId': playlist.id, 'videoId': video.id}) }}">Supprimer de cette playlist</a>
+										<a href="{{ path('playlist_remove_video', {'playlistId': playlist.id, 'videoId': video.id}) }}">Supprimer de cette playlist</a>
 									</li>
 								{% endfor %}
 							</ul>
@@ -36,7 +36,7 @@
 							<ul>
 								{% for playlist in playlistsExcVideo %}
 									<li>
-										<a href="{{ path('video_add_playlist', {'playlistId': playlist.id, 'videoId': video.id}) }}">{{ playlist.label }}</a>
+										<a href="{{ path('playlist_add_video', {'playlistId': playlist.id, 'videoId': video.id}) }}">{{ playlist.label }}</a>
 									</li>
 								{% endfor %}
 							</ul>
